+++ wp-subtitle.php
@@ -4,7 +4,7 @@
 Plugin Name: WP Subtitle
 Plugin URI: http://wordpress.org/plugins/wp-subtitle/
 Description: Adds a subtitle field to pages and posts. Possible to add support for custom post types.
-Version: 2.8.1
+Version: 2.9
 Author: Ben Huson, Husani Oakley
 Author URI: https://github.com/benhuson/wp-subtitle
 License: GPLv2
@@ -69,6 +69,33 @@ class WPSubtitle {
 	public static function _add_default_post_type_support() {
 		add_post_type_support( 'page', 'wps_subtitle' );
 		add_post_type_support( 'post', 'wps_subtitle' );
+		add_post_type_support( 'revision', 'wps_subtitle' );
+
+		add_filter( 'the_preview', array( 'WPSubtitle', 'the_preview' ), 10, 2 );
+
+	}
+
+	/**
+	 * Returns the autosaved data to make changes visible in preview mode.
+	 *
+	 * @since  2.9
+	 *
+	 * @param   object         $post   Post object.
+	 * @param   object         $query  Query object.
+	 * @return  WP_Post|false          The autosaved data or false on failure or when no autosave exists.
+	 */
+	public static function the_preview( $post, $query ) {
+
+		if ( isset( $_GET['preview_id'] ) ) {
+			return wp_get_post_autosave( $post->ID );
+		}
+
+		if ( $revisions = wp_get_post_revisions( $post->ID ) ) {
+			return array_shift( $revisions );
+		}
+
+		return $post;
+
 	}
 
 	/**
@@ -82,7 +109,7 @@ class WPSubtitle {
 		$post_types = (array) get_post_types( array(
 			'_builtin' => false
 		) );
-		$post_types = array_merge( $post_types, array( 'post', 'page' ) );
+		$post_types = array_merge( $post_types, array( 'post', 'page', 'revision' ) );
 		$supported = array();
 		foreach ( $post_types as $post_type ) {
 			if ( post_type_supports( $post_type, 'wps_subtitle' ) ) {